pyproject.toml
[project]
name = "ai-skill"
version = "0.1.0"
description = "AI agent for the complete academic research pipeline — from topic planning to publication."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "instructor>=1.6.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "pypdf>=4.3.0",
    "duckduckgo-search>=6.3.0",
    "exa-py>=1.0.0",
    "firecrawl-py>=1.0.0",
    "arxiv>=2.1.0",
    "pyyaml>=6.0.2",
    "rich>=13.8.0",
    "typer>=0.12.0",
    "anyio>=4.6.0",
    "lxml>=5.3.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "pytest-recording>=0.13.0",
    "respx>=0.21.0",
    "pyright>=1.1.390",
    "ruff>=0.6.0",
    "pre-commit>=4.0.0",
    "types-pyyaml",
    "types-lxml",
    "types-beautifulsoup4",
]
pdf-export = [
    "weasyprint>=62.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-mermaid2-plugin>=1.2.0",
]

[project.scripts]
ai-skill = "ai_skill.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_skill"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]
select = [
    "E", "W",
    "F",
    "I",
    "N",
    "D",
    "UP",
    "ANN",
    "S",
    "RUF",
]
ignore = [
    "D100", "D101", "D102", "D103", "D104", "D105", "D107",
    "ANN101", "ANN102",
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
include = ["src"]
venvPath = "."
venv = ".venv"
